+++ core/src/main/java/com/edwinyosua/core/data/remote/repository/GameListRepository.kt
@@ -1,31 +1,34 @@
-package com.edwinyosua.core.data.repository
+package com.edwinyosua.core.data.remote.repository
 
 import android.util.Log
 import com.edwinyosua.core.data.remote.network.ApiResponse
 import com.edwinyosua.core.data.remote.network.ApiService
-import com.edwinyosua.core.data.remote.network.ResultsItem
-import kotlinx.coroutines.Dispatchers
+import com.edwinyosua.core.domain.home.repository.IGameListRepository
+import com.edwinyosua.core.domain.home.mapper.toDomain
+import com.edwinyosua.core.domain.home.model.Games
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
 
-class GameListRepository(private val apiService: ApiService) {
+class GameListRepository(private val apiService: ApiService) : IGameListRepository {
 
-    fun getGameList(): Flow<ApiResponse<List<ResultsItem>>> = flow {
+    override fun getGameList(): Flow<ApiResponse<List<Games>>> = flow {
         try {
             emit(ApiResponse.Loading)
             val response = apiService.getGameList()
-            val gameList = response.results
+            val gameList = response.results.toDomain()
+
             if (gameList.isNotEmpty()) {
                 emit(ApiResponse.Success(gameList))
             }
+
             if (gameList.isEmpty()) {
                 emit(ApiResponse.Empty)
             }
         } catch (e: Exception) {
+            Log.e("GameListRepository", e.toString())
             e.printStackTrace()
             emit(ApiResponse.Error(e.message.toString()))
-            Log.e("GameListRepository", e.toString())
         }
-    }.flowOn(Dispatchers.IO)
+    }
+
 }
